requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy
Pillow
scipy>=0.17.0
Flask==2.0.1
opencv-python==4.5.2.54